+++ src/sdk/CHANGELOG.md
@@ -41,6 +41,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Basic Taxonomy support using the Graph API (pending work: relations, properties) #104 [jansenbe - Bert Jansen]
 - Support working with files #78 [ypcode - Yannick Plennevaux]
 - Use CSOM to create content types with a specified id #108 [ypcode - Yannick Plennevaux]
+- Support for making interactive SharePoint REST requests + upload of files #112 [jansenbe - Bert Jansen]
 
 ### Changed